requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
bert-score==0.3.13
# Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
certifi==2023.7.22
# cffi @ file:///croot/cffi_1670423208954/work
chardet==5.1.0
charset-normalizer==2.1.0
click==8.1.3
cmake==3.27.0
# colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.0.7
# cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
Cython==0.29.35
# dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
dataclasses-json==0.5.7
datasets==2.13.1
dill==0.3.6
evaluate==0.4.0
# filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1686612785345/work
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.6.0
gensim==4.3.1
greenlet==2.0.2
huggingface-hub==0.16.4
# idna @ file:///croot/idna_1666125576474/work
# importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==5.12.0
Jinja2==3.1.2
# joblib @ file:///croot/joblib_1685113087166/work
kiwisolver==1.4.4
langchain==0.0.181
lit==16.0.6
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mkl-fft==1.3.6
# mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
# multiprocess @ file:///croot/multiprocess_1668006441451/work
mypy-extensions==1.0.0
# neo4j-driver==5.9.0 ### 太舊了 以前是叫 Neo4j-driver / 現在叫 Neo4j
networkx==3.1
nltk==3.8.1
# numexpr @ file:///croot/numexpr_1683221822650/work
# numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.7
openapi-schema-pydantic==1.2.4
# packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
Pillow==9.5.0
# pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
protobuf==3.20.3
pyahocorasick==2.0.0
pyarrow==12.0.1
pycocoevalcap @ git+https://github.com/salaniz/pycocoevalcap.git@a24f74c408c918f1f4ec34e9514bc8a76ce41ffd
# pycocotools==2.0.6
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.8
# pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
# PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
# python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
# pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
# PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757097602/work
rank-bm25==0.2.2
# regex @ file:///tmp/abs_41f5bce5-0a2e-45aa-b231-1fd2fbd57753gfpe6sjm/croots/recipe/regex_1658257178822/work
# requests @ file:///croot/requests_1682607517574/work
responses==0.18.0
sacremoses==0.0.43
# scikit-learn @ file:///croot/scikit-learn_1684954695550/work
scipy==1.10.1
# six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
# smart-open @ file:///home/conda/feedstock_root/build_artifacts/smart_open_split_1673202927732/work/dist
SQLAlchemy==2.0.15
sympy==1.12
tenacity==8.2.2
# threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
# tokenizers==0.13.0.dev0
tokenizers
torch==2.0.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
# urllib3 @ file:///croot/urllib3_1680254681959/work
# xxhash @ file:///croot/python-xxhash_1667919521615/work
yarl==1.9.2
zipp==3.15.0
scikit-learn
langchain_community
python-dotenv
neo4j
groq
pyvis